+++ radicle-crypto/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "radicle-crypto"
+license = "MIT OR Apache-2.0"
+version = "0.1.0"
+authors = [
+  "Alexis Sellier <alexis@radicle.xyz>",
+  "Fintan Halpenny <fintan.halpenny@gmail.com>",
+]
+edition = "2021"
+
+[features]
+test = ["fastrand", "quickcheck"]
+ssh = ["base64", "radicle-ssh", "sha2"]
+
+[dependencies]
+ed25519-compact = { version = "1.0.12", features = ["pem"] }
+multibase = { version = "0.9.1" }
+serde = { version = "1", features = ["derive"] }
+sqlite = { version = "0.27.0", optional = true }
+thiserror = { version = "1" }
+
+[dependencies.fastrand]
+version = "1.8.0"
+default-features = false
+optional = true
+
+[dependencies.git-ref-format]
+version = "0.1"
+optional = true
+
+[dependencies.quickcheck]
+version = "1"
+default-features = false
+optional = true
+
+[dependencies.radicle-ssh]
+path = "../radicle-ssh"
+version = "0"
+default-features = false
+optional = true
+
+[dependencies.sha2]
+version = "0.10.2"
+optional = true
+
+[dependencies.base64]
+version = "0.13"
+optional = true
+
+
+[dev-dependencies]
+fastrand = { version = "1.8.0" }
+quickcheck_macros = { version = "1", default-features = false }
+quickcheck = { version = "1", default-features = false }
+zeroize = { version = "1.5.7" }
