+++ radicle-tools/src/rad-auth.rs
@@ -1,6 +1,6 @@
 fn main() -> anyhow::Result<()> {
     let keypair = radicle::crypto::KeyPair::generate();
-    radicle::ssh::agent::register(&keypair.sk)?;
+    radicle::crypto::ssh::agent::register(&keypair.sk)?;
 
     let profile = radicle::Profile::init(keypair)?;
