+++ radicle-tools/src/rad-self.rs
@@ -2,10 +2,10 @@ fn main() -> anyhow::Result<()> {
     let profile = radicle::Profile::load()?;
 
     println!("id: {}", profile.id());
-    println!("key: {}", radicle::ssh::fmt::key(profile.id()));
+    println!("key: {}", radicle::crypto::ssh::fmt::key(profile.id()));
     println!(
         "fingerprint: {}",
-        radicle::ssh::fmt::fingerprint(profile.id())
+        radicle::crypto::ssh::fmt::fingerprint(profile.id())
     );
     println!("home: {}", profile.home.display());
