+++ radicle/src/test.rs
@@ -2,5 +2,4 @@
 pub mod arbitrary;
 pub mod assert;
 pub mod fixtures;
-pub mod signer;
 pub mod storage;
